Split `Importer` into individual classes for each decl type.
This is one step towards a more query-based importer architecture, by
encapsulating the individual import methods and extracting an interface for
their interaction with the importer.

The extracted interfaces of the `DeclImporter` and `ImportContext` are not meant
to be permanent. They merely extract the status quo and provide a basis for
future improvements.

I also plan on moving individual `DeclImporter`s into their own files. Maybe
also the `Invocation` and the `ImportContext`. I haven't done that in this CL to
make review easier.

PiperOrigin-RevId: 443642590