Support custom path to crubit/rs_bindings_from_cc/support directory.
The generated thunks and other code in ..._rs_api_impl.cc might depend
on Crubit's support headers like so:

    #include "third_party/crubit/rs_bindings_from_cc/support/..."

Before this CL, `src_code_gen.rs` would hardcode the path to the support
directory.  After this CL, this is replaced with an argument of a new
cmdline parameter: --crubit_support_path.

PiperOrigin-RevId: 447003312