Qualify `detail` and `rust_std` where needed
the `detail` module is now top level and as such we need the `crate` qualifier when calling thunks from impls. For `rust_std` we used to have `use super::*` in each namespace, but that doesn't cut it for nested namespaces, as for those we'd need a `pub use super::*`. Instead I went with the `crate` qualifier with the hope that we'll find time at some point to fix b/227790881

PiperOrigin-RevId: 448207422