Support unnamed fields.
This is not a very important case, but it does occur (e.g., in `<bits/timex.h>`).

The reason to support this is to get rid of one of the remaining cases where we fail to import fields. If we could eliminate that fully, the overall importer setup could be simplified.

PiperOrigin-RevId: 448246613