Use `#[ctor::recursively_pinned]` in interop tooling.
That brings a lot with it:

1. `Drop` is now `PinnedDrop`, which mostly removes special-casing for destructors.
2. The `ctor!()` macro can now be used to construct (aggregate...) structs.
3. Pin projection can be used to set fields in pinned structs.
4. The `impl !Unpin` is now implicit.

While #1 turns out to be the most visible change in this CL, the actual reason for this CL is #2 and #3. Without this CL, `!Unpin` aggregate structs can't be used from Rust in any useful way. (We can default-construct them, but not set any of the fields.)

The tests for #1 already existed, just needing to be reworded. Tests have been added for #2 and #3.

PiperOrigin-RevId: 449076947