Add test for converting lifetime names to tokens.
Worth, uh, mentioning just obliquely that the ToTokens implementation can fail. (Was true of the old code too.) There's no validation coming in from the IR that the name string doesn't e.g. have quotes or spaces in it, etc.

At some point we will want to tighten up the IR, but not today.

PiperOrigin-RevId: 451363614