Merge `TypeMapper` back into `Importer` - part 1/N: `Contains`.
This CL is one of steps toward merging `TypeMapper` back into
`Importer`.  The merge is:
- possible (after https://github.com/google/crubit/commit/16a5610d96eda5720bb57b5f927b00f00635ddf0 which means that there is only 1 instance
  of `TypeMapper`, because `ImportFields` doesn't anymore need to
  temporarily assume that the containing record can be imported),
- desirable (to support b/228868369 we need to let `ConvertType` call
  into `Importer::ConvertTemplateSpecializationType`).

This CL makes `TypeMapper::Contains` a private method of `TypeMapper`
and exposes this private method to `Importer` via `friend` declaration
(`TypeMapper` and the `friend` declaration will disappear once the merge
is complete).

Callers of `TypeMapper::Contains` are transitioned to call the newly
introduced, equivalent `Importer::HasBeenAlreadySuccessfullyImported`.

PiperOrigin-RevId: 451503758