Automatically determine which lifetimes to declare in `impl` vs `fn`.
The problem with `declare_lifetimes` is that it's all-or-nothing -- either everything is a lifetime parameter of the `impl`, or of the `fn`. And so `declare_lifetimes: true` works for static functions, but not instance functions where _some_ lifetime parameters are on the `fn` (e.g. for `&self`), and others on the trait.

To do this requires structuring the trait names: rather than an opaque bag of tokens, it is a name + parameters. This feels like a good change, structure is good! And it allows us to do things like automatically determine which lifetimes must be parameters on the `impl` block. :)

PiperOrigin-RevId: 452617483