Emit `private` and `protected` fields as opaque blobs of bytes.
Before this CL, fields with a type that is unsupported Crubit would be
emitted as a private field that is backed by an opaque blob of bytes
(represented by a Rust array `[u8; N]` of appropriate size).

After this CL, `private` and `protected` fields with a Crubit-supported type
are also emitted as an opaque blob of bytes.

This CL is desirable as a performance optimization - e.g. to avoid
generating bindings for template instantiations that are only ever used
in non-public fields.

PiperOrigin-RevId: 452627863