Generate IR for namespace items coming from a different target
In the past we didn't have unique identifiers for the IR items, and similar items that appeared in multiple targets (eg reopened namespace accross targets) would confuse the bindings generator. We're past that point.
Now we try to import all the decls from the translation unit, but we only put top level items from the current target into ir.top_level_items, so that we still generate source code only for the current target.

PiperOrigin-RevId: 453523844