Generate bindings for inline namespaces
In the bindings, we use fully qualified identifiers everywhere.

In the following example, `C` will always be referred to in the bindings as `A::B::C` regardless of whether the C++ equivalent was `A::B::C`, `A::C` or `C`.
```
namespace A {
inline namespace B {
struct C{};
}  // B
}  // A
```
So, we don't need any special logic for inline namespaces.

PiperOrigin-RevId: 453906326