Various fixes to `TypeLoc` version of `GetTemplateArgs()`.
- It's not always possible to retrieve `TypeLoc`s for the template arguments,
  so the return value is now wrapped in a `std::optional` (see comments in code
  for details).

- We should handle `RecordTypeLoc` in the same way that we handle `RecordType`
  in the `Type` version of `GetTemplateArgs()`.

These fixes prevent breakage once the next CL starts using this function in
lifetime_annotations. This change doesn't add any unit tests because it's hard
to test this behavior in isolation, but the upcoming change will exercise all of
the changes here. I chose to separate the two changes to make review easier.

PiperOrigin-RevId: 456702187