Thread `TypeLoc` through `FunctionLifetimes` and `ValueLifetimes`.
A followup CL will extract `annotate_type` attributes from the `TypeLoc`.

Note: One might ask why all this mechanism is necessary. Couldn't the attribute
arguments be placed on the `AttributedType`, so that we could extract them
without having to access the `TypeLoc`?

However, putting the attribute arguments on the `AttributedType` would violate
the conceptual distinction between `Type` and `TypeLoc`. The attribute arguments
are `Expr`s, which are associated with a specific location in the source code.
In contrast, a `Type` is intended to be independent of source location.

PiperOrigin-RevId: 456703847