Add `CreateThisLifetimes()` to `FunctionLifetimeFactory` interface.
This is needed because we don't have a `TypeLoc` for the `this` parameter, as it
is never spelled out explicitly in the code. Instead, we directly pass in the
lifetime name for the `this` pointer. (This CL is only changing the interface
and therefore passing in null for the time being; the next CL will actually pass
in the lifetime name.)

PiperOrigin-RevId: 456706667