Clarify comment on passing a null `TypeLoc` in the case of parameter packs.
I've now thought about this some more and realized that passing a null `TypeLoc`
is the right thing to do here; see changed comment for details.

This change also adds a test that triggers this code path (verified by
temporarily adding an `assert(false)`). It turns out that this test triggers
another minor issue -- see TODO in the code.

PiperOrigin-RevId: 457415113