Use `Rc` for items in the `IR`, and `RsTypeKind`.
This way, we no longer need to keep track of `IR` lifetime throughout every type. Additionally, it becomes possible to migrate `RsTypeKind` into `IR` itself, or -- as I intend to do in a followup CL -- inside Salsa, which cannot use references, as it doesn't allow the salsa db to have lifetime parameters.

`Rc` copies are hella cheap, so this seems nice. Even if the performance cost were larger (for example, `Arc` instead of `Rc`), we actually do very few copies, so we should expect this to remain not-a-big-deal forever. The only copy we should expect is when we actually create the `RsTypeKinds`, which was already expensive and should (ideally) only happen once. (It currently can happen a lot, but...)

PiperOrigin-RevId: 458112800