Use salsa in src_code_gen.rs.
This only actually changes generate_func into a query for now -- notice that it's retrieved in two places, so this is actually saving work!

More importantly, this shows the shape of the change. Any given function can be migrated to a query by changing its parameter from a `&mut Database` to a `&dyn BindingsGenerator`.

A quick note on `PtrEq` -- this used to be possible without this; discussion at https://salsa.zulipchat.com/#narrow/stream/145099-general/topic/Return.20values.20that.20aren't.20comparable/near/286482700.

PiperOrigin-RevId: 458114855