Refactor `AnalyzeSingleFunctionBody()`.
Now that we can analyze defaulted functions, the name of
`AnalyzeSingleFunctionBody()` is no longer appropriate, as it also analyzes
functions without a body.

Rename to `AnalyzeSingleFunction()` and factor out a function
`AnalyzeFunctionBody()` that deals only with function bodies.

An additional effect is that we now do additional things for defaulted functions
that we used to do only for functions with bodies, namely emitting debug info
and, importantly, propagating static lifetimes to pointees. (Strictly speaking,
this makes this change more than a refactor.)

PiperOrigin-RevId: 458209309