Introduce the ObjectRepository::CreateObject...() object allocation functions.
Replace the Object::Create...() functions with corresponding constructors; these
should, however, only be used in tests where an ObjectRepository is not
available.

Everywhere else, we replace calls to Object::Create...() with the new
ObjectRepository::CreateObject...() functions. However, at the callsites, we
immediately convert back to an Object value. This is fine because Object still
has value semantics, but it will allows to gradually start replacing Object
values with `const Object *` pointers.
PiperOrigin-RevId: 458412874