Use `&Database`, not `&mut Database`.
This was a slight confusion on my part. If I understand it correctly, `Database` actually implements `BindingsGenerator`, meaning all these functions are callable from const references. (The caching is internal mutability.)

The other change I could make is to have all of these accept `&dyn BindingsGenerator`, but for now I'm only doing that where necessary: where the function is actually implemented as a salsa query, or where it might be called from a salsa query.

PiperOrigin-RevId: 459493349