Prevent the direct creation of non-Unpin structs, without the `ctor!()` macro.
For migratability from C++, we want to be able to allow people to create non-trivially-relocatable structs using the `ctor!()` macro -- much like how in C++, you can do `S {.x=3}`, in Rust, you should be able to do `ctor!(S {x: 3})`. However, we also need to _forbid_ the creation of the struct (in safe code) via normal `S {x: 3}` syntax. Being non-trivially-relocatable, the very moment you create such a struct you've likely created an object whose state is invalid and which is UB to touch!

So we split the struct into two parts: the actual type in normal use is specially marked so that you can't directly create it (in this case, via `#[non_exhaustive]`, but a private field would also work). And then there is a second struct with the same layout which is used for initialization in the `ctor!()` macro. (We're effectively transmuting between the two.) This means that the only way to create such structs in safe code is via the `ctor!()` macro, and our abstractions become sound again.

This is the followup to https://github.com/google/crubit/commit/a330af6d6a3b1c803c9c6f5f67b8888d92e7f28e and, I believe, closes the final remaining soundness hole in `ctor.rs`.

PiperOrigin-RevId: 462276896