Allow either order of `derive(CtorFrom_Default)` / `#[recursively_pinned]`.
Primary error: the layout of `CtorInitializedFields` is ignored, I can just clear the attrs! Oopsie woopsie.

However, even if we do that, there's other problems that make these macros compose very poorly in the original order. Apparently, ideally, either order should work, but this is definitely messy to support.

Another solution would be to detect if they're misordered and swap the order in the macro.

I think this is worth deferring to some later time to do "properly" -- this only matters during the process of porting a C++ class to Rust, and that's somewhat of a theoretical concern given the current state of interop.

PiperOrigin-RevId: 463065300