Only emit a C++ thunk if there was Rust code generated for the same function.
A careful observer would note that it's still possible that a function _was_ generated, but that it doesn't call the C++ thunk, however, this is meant to be captured today by the "out of band" communication channel of `can_skip_cc_thunk`.

I think that, now that these can directly call each other, maybe we should retire `can_skip_cc_thunk` and instead feature this information in the return value of `generate_func`. But for now... fixed!

---

The fix itself is essentially two lines of code, but the change breaks many tests, which make assertions about the generated C++ thunks. If we remove the thunks, the tests fail! Where I was able to, I tried to keep (a version) of the test by fixing the C++ header so that it would emit functions. Where I didn't know how to do this (in particular, test_implicit_template_specializations_are_sorted_by_mangled_name), I deleted the broken assertions. (They did not seem critical to me, and the tests still otherwise are working and testing things.)

PiperOrigin-RevId: 463097005