Implement C++17 guaranteed copy elision for return values.
Actually, I'd thought I did that, but, as they say, "perfect forwarding isn't". By using construct_at, we were materializing the prvalue to bind to an rvalue reference to invoke the move constructor! This is a place where construct_at is not equivalent to placement new. In doing this, we were not taking advantage of guaranteed copy elision.

This also adds an actual test for the feature. Honestly it's really cool to see written out, in Rust, a use of C++ guaranteed copy elision:

```c++
ctor::emplace! {
    let x = ReturnsNonmovable();
}
assert_eq!(x.addr, &*x as *const _ as usize);
```
PiperOrigin-RevId: 463579372