Implement non-`Unpin` function parameters for non-trait functions/methods.
This allows Rust code to pass non-trivially-relocatable objects (and other non-`Unpin` objects) by value to C++, using C++11/14 move semantics.

C++17 guaranteed copy elision cannot be implemented until after we have the "reverse" direction of C++ interop: effectively, we need to convert a `Ctor<Output=T>` into a C++ function object which returns `T` by value. This cannot be done in an ABI-agnostic way: it must take advantage of knowledge of the platform and how it implements return slots. And so it's a pretty big task. Given that it's definitely possible ("just" a matter of gluing together ABIs), and not yet pervasively required, this ought to be safe for now.

(We do still have guaranteed copy elision for return values, which is somewhat more useful.)

---

#### TODO: Traits

This change does not implement pass by value in traits, where the parameter type can't just be replaced by an `impl` because that introduces a type parameter.

(This breaks all traits, not just e.g. `Assign`. However, the only place this shows up in the goldens is the `Assign` trait.)

The core problem is that in a trait, you can't just drop an `impl Ctor` in there: it's an extra type parameter.
This CL would have, naively, created something like the following if it weren't specifically carved out as todo:

```rs
impl Assign<S> for S {
  fn assign(&mut self, x: impl Ctor<Output=S> { ... }
}
```

When in fact the correct output is:

```rs
impl<__T> Assign<__T> for S where
    __T : Ctor<Output=S> {
  fn assign(&mut self, x: __T) { ... }
}
```

However, correctly implementing that is best left to a separate change, because it's a bit of work to thread this through appropriately.

(In theory, we need to handle both trait type parameters and associated types -- but in practice I suspect only the former at least for now.)

PiperOrigin-RevId: 463585180