Crash if our hardcoded function IDs fail to parse.
This is always a bug in the interop tooling, _not_ the code. And unfortunately, the error messages are worthless: all you see is "unexpected token" without an explanation of what that means. On a crash, however, it will point at the line number this failed at, which can give a hint as to what mistake you made. :)

PiperOrigin-RevId: 463585477