Cleanup: delete `maybe_first_api_param`, use `param_types.get(0)` instead.
`maybe_first_api_param` is used to change the `self` parameter -- for `Clone` in particular, the `this` argument of the C++ function becomes the _return value_ of the Rust function. And so we remove the first parameter and use the _argument_ as `self`.

However it's basically unnecessary. The control flow is difficult to track here, but everywhere else we actually mutate the vectors of parameters/arguments/etc. -- and we can do the same for the parameter types, as well. This makes our code simpler and more obviously correct. For example, I think this newly-updated comment makes more sense and is easier to follow along than the previous version: `// If param_types[0] exists, so do api_params[0] and thunk_args[0].`

This change may at first seem like it changes behavior where `param_types` is used, but it's actually only used after this mutation to pull out the `this` pointer -- where the code actually should've been using `maybe_first_api_param` anyway. The fact that it _wasn't_ is a bug, fixed by this CL. (And the fact that it got in and looked fine -- in fact, I was the code reviewer that let it in -- shows why we should delete `maybe_first_api_param`!)

Well, that's a smidgen of a lie. In fact, there's no way to trigger this latent bug: those code paths are only hit when `maybe_first_api_param` _isn't_ relocated to type[1], and so there is no behavior change at all. But if there were a behavior change, it would've been a bug fix. :)

---

(The context for this change is actually that I'm looking to see if some of this logic can't be pulled out to a separate function. Removing unnecessary local variables/state is step #1!)

PiperOrigin-RevId: 464613456