Move function signature handling to a separate function.
I think this is a good call no matter what -- it's just a lot of code smeared around.

I also *might* need this to generate a new blanket trait impl for handling `Ctor` in trait parameters -- because the argument handling code wasn't a function, I couldn't invoke it. Now that it is a function, I think I can. (Maybe with some more tweaks to the interface, of course.)

PiperOrigin-RevId: 464616745