Implement bindings from `operator+=` to `AddAssign`.
Because Rust compound assignment expressions produce unit, these bindings discard the operator's return value. And because `AddAssign::add_assign` receives `&mut self`, bindings are not generated for const left-hand sides.

An idiomatic `operator+=` member function that takes the right-hand side as a reference and returns a reference binds correctly. But there is no lifetime information on the first parameter when binding an equivalent friend or free function, so no bindings are generated for this common case. A similar friend or free function that takes the right-hand side by value or does not return a reference does work as expected.

PiperOrigin-RevId: 465169729