Use a `mov!` macro, and delete the `Move<P>` type. (Same for "const moves".)
This nearly completely finishes implementing move semantics for C++.
Where, before, we had not finished hooking together `Move<P>` with the
`RvalueReference<T>` in bindings, it now Just Works. Correspondingly, I've
added a handful of direct calls to `mov!(x)` to tests, which would not
have compiled before this change.

This also means that the linked bug is effectively done: the only remaining significant work item that I'm aware of is making it accept `Ctor` on the inside, which is optional.

The newly introduced `mov!` macro itself is very simple, and nearly
identical to C++'s `std::move` in what it does. The abbreviated form is:

```rs
macro_rules! mov {
    ($p:expr) => {
        RvalueReference( {$p}.as_mut() )
    };
}
```

In other words, it rust-moves its argument (which is a `Pin<P>`) to a temporary,
and then evaluates to a `RvalueReference<T>` pointing at the pointee of that temporary.

Just like the original mov function, this means that you can automatically prevent
use-after-move if you pass by value:

```rs
foo(mov!(x));
foo(mov!(x)); // fails to compile
```
... but if you call `.as_mut()` you can also choose to use-after-move on purpose:

```rs
foo(mov!(x.as_mut())); // moves x
// `x` still exists, but refers to a moved-from object.
foo(mov!(x));  // succeeds
```

---

This approach is better in nearly every way:

1.  By having exactly one rvalue reference type, not two families of types, we eliminate trait implementation issues: we have a single monomorphized type, which is easy to deal with when defining traits which accept rvalue references.

    Consider the naive trait implementation for `From` that supports two different rvalue reference types. This is broken:

    ```rs
    impl<P: DerefMut<Target=A>> From<Move<P>> for S {...}
    impl<P: DerefMut<Target=B>> From<Move<P>> for S {...}
    ```

    Correctly designing traits that deal with Move is a challenge, and an apparently unnecessary one.

2.  By having exactly one rvalue reference type, we simplify the model. There's just plain less code now. `mov` works in the "obvious" way, and is easy to explain.

The downsides:

1.  `RvalueReference` is now a strange type (e.g. `!Unpin`). I think this is OK, and anyway I think we should work on fixes to this that make it behave in line with `&T`. (In particular, I believe we should stop using `Unpin` as a `SelfCtor` trait, and define a new auto trait for it... but that's the future, and this is the present, and I still think this is an acceptable loss in the here and now.)

2.  More macros. (I believe this is more than outweighed by the simplicity gain.)

3.  It's no longer possible to store a `Move<P>` durably, e.g. a `Move<Box<T>>`.
    However, since such APIs don't exist in C++ anyway, this is no major loss.

---

### Why wasn't it done this way from the start?

This really is the "obvious" implementation in many ways. But I didn't realize you
could do this at the time, only now that I am revisiting move semantics for
implementing them on traits.

When I originally wrote `mov()`, I thought about making it a macro, but the form I had thought of was something like:

```rs
($p:expr) => {RvalueReference($p.as_mut())};
```

... which would not have disallowed use-after-move. I had thought, at the time, that adding a move here would break callers and not work:

```rs
($p:expr) => {RvalueReference({$p}.as_mut())};
```

... but it actually works just fine. You can store the object as a temporary, which lives for the duration of the enclosing statement. This means that mov!(x) is fine as a parameter. You just can't do something like `let x = mov!(y);`, because the lifetime of the temporary will end and you can't use `x`.

(I learned this "make a temporary, then take a mut reference to it" trick when implementing the `emplace!(x)` expression macro, which does exactly this to materialize temporaries.)

This behavior up exactly with C++, though for completely different reasons. In C++, the return value of `std::move()` must be used in the current statement in order to retain rvalue status. If you assign it to a variable, it becomes an lvalue, and you'd need to move it again. So people used to coding in existing C++ patterns will not feel weirded out by the shape in which one uses the Rust `mov!` macro.

### Example use-after-move-error

Just to show this really does work, I added a `mov!()` by value in a test to see how it explodes. As so:

```
error[E0382]: use of moved value: `x`
    --> third_party/crubit/rs_bindings_from_cc/support/ctor.rs:1624:18
     |
1621 |         let x: Pin<Box<S>> = Box::pin(S);
     |             - move occurs because `x` has type `std::pin::Pin<Box<S>>`, which does not implement the `Copy` trait
1622 |         fn takes_rvalue_reference(_: RvalueReference<S>) {}
1623 |         takes_rvalue_reference(mov!(x));
     |                                     - value moved here
1624 |         let _x = x; // fails to compile: x is moved!
     |                  ^ value used here after move

error: aborting due to previous error
```
PiperOrigin-RevId: 465265577